Split isdefs
