+++ lookslike/isdefs/duration.go
@@ -1,0 +1,21 @@
+package isdefs
+
+import (
+	"fmt"
+	"github.com/elastic/lookslike/lookslike/paths"
+	"github.com/elastic/lookslike/lookslike/results"
+	"time"
+)
+
+// IsDuration tests that the given value is a duration.
+var IsDuration = Is("is a duration", func(path paths.Path, v interface{}) *results.Results {
+	if _, ok := v.(time.Duration); ok {
+		return results.ValidResult(path)
+	}
+	return results.SimpleResult(
+		path,
+		false,
+		fmt.Sprintf("Expected a time.duration, got '%v' which is a %T", v, v),
+	)
+})
+
