+++ lookslike/isdefs/duration_test.go
@@ -1,0 +1,13 @@
+package isdefs
+
+import (
+	"testing"
+	"time"
+)
+
+func TestIsDuration(t *testing.T) {
+	id := IsDuration
+
+	assertIsDefValid(t, id, time.Duration(1))
+	assertIsDefInvalid(t, id, "foo")
+}
